FarmerApp.Api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
tensorflow==2.19.0
pillow==10.1.0
numpy>=1.26.0